+++ tests/test_node.py
@@ -1,9 +1,3 @@
-import pytest
-
-from ididi.errors import (  # GenericDependencyNotSupportedError,; MissingAnnotationError,
-    MissingReturnTypeError,
-    UnsolvableDependencyError,
-)
 from ididi.node import DependentNode